requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.4
dj-database-url==0.5.0
Django==3.0.8
django-fernet-fields==0.6
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
